% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb_ms2}
\alias{hmdb_ms2}
\title{Human Metabolome Database}
\format{
A databaseClass object.
}
\usage{
hmdb_ms2
}
\description{
HMDB
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
\keyword{datasets}
