% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{massbank_ms2}
\alias{massbank_ms2}
\title{massbank database database}
\format{
A databaseClass object.
}
\usage{
massbank_ms2
}
\description{
massbank database database
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
\keyword{datasets}
