% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mona_ms2}
\alias{mona_ms2}
\title{MoNA - MassBank of North America database}
\format{
A databaseClass object.
}
\usage{
mona_ms2
}
\description{
MassBank of North America database
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
\keyword{datasets}
